+++ packages/state_beacon_core/lib/src/extensions/chain.dart
